+++ ydb/core/kqp/provider/yql_kikimr_gateway.h
@@ -527,6 +527,7 @@ struct TTtlSettings {
     struct TTier {
         TDuration ApplyAfter;
         std::optional<TString> StorageName;
+        std::optional<TString> ObjectKeyPrefix;
     };
 
     TString ColumnName;