+++ ydb/core/ydb_convert/table_settings.cpp
@@ -715,6 +715,10 @@ bool FillPublicTtlSettingsImpl(Ydb::Table::TtlSettings& out, const TTtl& in, Ydb
                 break;
             case NKikimrSchemeOp::TTTLSettings::TTier::ActionCase::kEvictToExternalStorage:
                 outTier.mutable_evict_to_external_storage()->set_storage(inTier.GetEvictToExternalStorage().GetStorage());
+                if (inTier.GetEvictToExternalStorage().HasObjectKeyPrefix()) {
+                    outTier.mutable_evict_to_external_storage()->set_object_key_prefix(inTier.GetEvictToExternalStorage().GetObjectKeyPrefix());
+                }
+
                 break;
             case NKikimrSchemeOp::TTTLSettings::TTier::ActionCase::ACTION_NOT_SET:
                 return bad_request("Undefined tier action");
@@ -841,6 +845,10 @@ bool FillSchemeTtlSettingsImpl(TTtl& out, const Ydb::Table::TtlSettings& in, Ydb
                         break;
                     case Ydb::Table::TtlTier::kEvictToExternalStorage:
                         outTier->MutableEvictToExternalStorage()->SetStorage(inTier.evict_to_external_storage().storage());
+                        if (inTier.evict_to_external_storage().has_object_key_prefix()) {
+                            outTier->MutableEvictToExternalStorage()->SetObjectKeyPrefix(inTier.evict_to_external_storage().object_key_prefix());
+                        }
+
                         break;
                     case Ydb::Table::TtlTier::ACTION_NOT_SET:
                         return bad_request("Tier action is undefined");